+++ WeatherProjectTestOrion/Main/View/TestPluginView/TestSwiftUIPluginForFigma/DetailsForCurrentWeatherViewTP.swift
